style : 예전버전으로 잘못 올린 것 수정

+++ src/public/js/reviewView.js
@@ -1,6 +1,5 @@
 $(document).ready(function () {
     const reviewId = new URLSearchParams(location.search).get('reviewId');
-    console.log(reviewId)
     getReviews(reviewId);
 });
 
@@ -11,8 +10,7 @@ function getReviews(reviewId) {
     })
         .then((res) => {
             let { id, title, content, createdAt, image, user } = res.data;
-            console.log(res.data)
-
+            let userId = user.id
             const createdTime = new Date(createdAt);
             const year = createdTime.getFullYear();
             const month = createdTime.getMonth() + 1;
@@ -25,7 +23,28 @@ function getReviews(reviewId) {
             if (minute.toString().length === 1) {
                 minute = '0' + minute.toString();
             }
-            let temp = `<div class="title">${title}</div>
+            if (image === "") {
+                let temp = `<div class="title">${title}</div>
+                <div class="info">
+                    <dl>
+                        <dt>글쓴이</dt>
+                        <dd>${user.nickname}</dd>
+                    </dl>
+                    <dl>
+                        <dt>작성일</dt>
+                        <dd>${year}년 ${month}월 ${day}일 ${hour}시 ${minute}분</dd>
+                    </dl>
+                </div>
+                <div class="profile-image">
+                </div>
+                <div id="content">${content}</div>
+                
+                <div class="btWrap">
+                    <div class="buttons2"></div>
+                </div>`
+            $('.boardView').append(temp);
+            } else {
+                let temp = `<div class="title">${title}</div>
                 <div class="info">
                     <dl>
                         <dt>글쓴이</dt>
@@ -42,15 +61,32 @@ function getReviews(reviewId) {
                 <div id="content">${content}</div>
                 
                 <div class="btWrap">
-                    <a href="/review/edit?reviewId=${id}" class="on">수정</a>
-                    <a onclick="deleteReview()" class="off">삭제</a>
-                    <a href="/place/placelist" class="on">목록</a>
+                    <div class="buttons2"></div>
                 </div>`
             $('.boardView').append(temp);
-
+            }
+            
+            loginUser3(userId);
         })
         .catch((err) => {
             console.log(err)
+        })
+}
+
+function loginUser3(userId) {
+    const id = new URLSearchParams(location.search).get('reviewId');
+    axios.get(`/auth/isLoggined`)
+        .then((res) => {
+            if (userId == res.data.id) {
+                let temp = `<a href="/review/edit?reviewId=${id}" class="on">수정</a>
+                            <a onclick="deleteReview()" class="off">삭제</a>
+                            <a href="/place/placelist" class="on">목록</a>`
+                $('.buttons2').append(temp)
+            }
+            else {
+                let temp = `<a href="/place/placelist" class="on">목록</a>`
+                $('.buttons2').append(temp)
+            }
         })
 }
 
@@ -61,8 +97,7 @@ function deleteReview() {
         method: 'delete',
     })
         .then((res) => {
-
-            confirm('삭제하시겠습니까?');
+            alert('삭제 완료!')
             window.location.href = '/place/placelist';
         })
         .catch((err) => {
